refactor: looks much better and understandable
