+++ src/db/words copy.json
@@ -1,5 +1,7 @@
 {
   "cat": { "en": "cat", "ru": "кот" },
   "dog": { "en": "dog", "ru": "собака" },
-  "ship": { "en": "ship", "ru": "корабль" }
+  "ship": { "en": "ship", "ru": "корабль" },
+  "sheep": { "en": "sheep", "ru": "овца" },
+  "cup": { "en": "cup", "ru": "чашка" }
 }
