+++ src/front/js/controller/state.controller.js
@@ -1,0 +1,10 @@
+export const renderState = {
+  server: {
+    toMainPageLink: "/",
+    toGamePageLink: "/game",
+  },
+  storage: {
+    toMainPageLink: "../index.html",
+    toGamePageLink: "./pages/game.html",
+  },
+};
