+++ src/front/js/helper/enum.js
@@ -1,0 +1,4 @@
+export const ENUM_SRC = {
+  server: "server",
+  local: "local",
+};
